+++ NEWS.md
@@ -1,7 +1,7 @@
 # r2dii.analysis (development version)
 
 * `target_market_share()` now correctly outputs `technology_share` with
-  multiple loans to the same company (@georgeharris2deg #262).
+  multiple loans to the same company (@georgeharris2deg #262, @ab-bbva #265).
 
 # r2dii.analysis 0.1.4